removed the Object module from protocol, that means that now no more custom signals, we use a simple module to use the publish/suscribe pattern, see the code where Signal is used, it's pretty simple and doesn't get into the code.
with that removal I removed all the existing code for menues and reimplemented it from scratch with plugins in mind, to allow to extend the menus by plugins. There is still work to do there but the foundations are in place.
 I also implemented the conversation toolbar, it doesn't do anything yet but it's a start.With the new refactoring the handling of the menus is in a separated place from the model.

Also fixed some bugs on ContactList when no item is selected, moved stock from protocol to gui (that is the correct place).

maybe more things...