Fixed my previous unstripped lines and a bug
A "fix" was done to the extension list refresh window, where if the refresh is taking too long, it would block the entire application.
However, the current fix doesn't actually stop the connection, it just closes the refresh dialog, which is causing the application to be blocked.
Also, if one clicks on the Download button and if their network sucks (like my current one) and the download doesn't go through, the application will freeze. This should be fixed by using a seperate thread to download. Bonus: ability to terminate that thread.
Added some todos that I want to do but can't as I need to figure out how some of the code works.