Redesigning plugin base, reworking plugin manager
Redesigned the PluginBase with ABC, which will work much nicer than the
current model. Also added a configurable property. This will eventually
allow the preference button to be disabled when a plugin is not
configurable.

This will break a lot of the current code due to the introduction of the
configurable property. This, however, should be easily fixable.