Added disabling the preference button.
Changes needs to be made to all supported plugins. I committed a test
plugin to my fork of emesene-supported-plugins in the experimental
branch.

Another thing I changed is from Plugin._description to
Plugin.description. I don't understand why we gave it _ prefix if it is
suppose to be public... Or am I wrong about this?