Re-added backwards compatibility
In my opinion, we should still use abc for PluginBase (It is a base
class, as the docstring suggests). Also, some existing plugins _will_
need to add the configurable() method in their code in order for the
Preference button to be usable.

Also, I personally think that the plugins should have a metadata file
instead of storing all the information in code. After all, they are
metadata, not real code.