papyon/switchboard: don't automatically close switchboards after 1 minute of inactivity
this tries to debug all remaining messages not sent/received issues.
Please note that the reason we had this check is because sometimes sockets get closed and we know nothing about them so we keep sending messages on a dead socket (bad issue)
However if we can instead debug that issue that could actually be a step in the right direction.