e3.common: Use the GitHub api v3 for downloading files from GitHub
This should make sure that we don't surpass the maximum file download limit,
because it is set to 5000 in the new api. This means no more corrupt files.
It seems to be a bit slower however, but that's most likely because we have
to download more data and convert it back from a base64 encoding.

Closes #970